csv publishing: Use display column in case of references
When fetching the tables from Grist, the values of reference columns are
just row IDs from other tables.
Instead let's switch to the display column, which is provided for
references.

Also remove the "L" prefix in arrays, which are just a technical piece
of information from Grist (meaning it is a list).